requirements.txt
# Development dependencies
black
flake8

# Package dependencies
click
dataclass-wizard
humanize
PyYAML
rich
sh >= 2.0
typer
